requirements.txt
lancedb
openai
pyarrow
pydantic
pyyaml
pyaml-env
